services/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.12.0
youtube-transcript-api==0.6.2
langchain==0.1.6
langchain-google-genai==0.0.11
langchain-core==0.1.23
faiss-cpu==1.7.4
